Move generate_sounds methods to voice.py
This fixes exception when running bot due to circular import.
